requirements.txt
langchain
streamlit
langchain_community
openai
python-dotenv
